+++ src/expectEvent.js
@@ -177,8 +177,10 @@ expectEvent.inLogs = deprecate(inLogs, 'expectEvent.inLogs() is deprecated. Use 
 expectEvent.inConstruction = inConstruction;
 expectEvent.inTransaction = inTransaction;
 
-expectEvent.not = {};
 expectEvent.notEmitted = notExpectEvent;
-expectEvent.not.inConstruction = notInConstruction;
-expectEvent.not.inTransaction = notInTransaction;
+expectEvent.notEmitted.inConstruction = notInConstruction;
+expectEvent.notEmitted.inTransaction = notInTransaction;
+
+expectEvent.not = deprecate(inLogs, 'expectEvent.not is deprecated. Use expectEvent.notEmitted instead.');
+
 module.exports = expectEvent;